Address comments and linter fixes
