--- pkg/command/plugin_bundle_download.go
@@ -1,40 +1,0 @@
-// Copyright 2023 VMware, Inc. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-package command
-
-import (
-	"github.com/spf13/cobra"
-
-	"github.com/vmware-tanzu/tanzu-cli/pkg/airgapped"
-	"github.com/vmware-tanzu/tanzu-cli/pkg/carvelhelpers"
-)
-
-type downloadPluginBundleOptions struct {
-	pluginDiscoveryOCIImage string
-	tarFile                 string
-}
-
-var dpbo downloadPluginBundleOptions
-
-func newDownloadBundlePluginCmd() *cobra.Command {
-	var downloadBundleCmd = &cobra.Command{
-		Use:   "download-bundle",
-		Short: "Download plugin bundle to the local system",
-		Long:  "Download plugin bundle to the local system",
-		RunE: func(cmd *cobra.Command, args []string) error {
-			options := airgapped.DownloadPluginBundleOptions{
-				PluginInventoryImage: dpbo.pluginDiscoveryOCIImage,
-				ToTar:                dpbo.tarFile,
-				ImageProcessor:       carvelhelpers.NewImageOperationsImpl(),
-			}
-			return options.DownloadPluginBundle()
-		},
-	}
-
-	f := downloadBundleCmd.Flags()
-	f.StringVarP(&dpbo.pluginDiscoveryOCIImage, "image", "", "", "plugin inventory image")
-	f.StringVarP(&dpbo.tarFile, "to-tar", "", "", "local tar file path to store the plugins images")
-
-	return downloadBundleCmd
-}
