--- pkg/command/plugin_bundle_upload.go
@@ -1,40 +1,0 @@
-// Copyright 2023 VMware, Inc. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-package command
-
-import (
-	"github.com/spf13/cobra"
-
-	"github.com/vmware-tanzu/tanzu-cli/pkg/airgapped"
-	"github.com/vmware-tanzu/tanzu-cli/pkg/carvelhelpers"
-)
-
-type uploadPluginBundleOptions struct {
-	sourceTar       string
-	destinationRepo string
-}
-
-var upbo uploadPluginBundleOptions
-
-func newUploadBundlePluginCmd() *cobra.Command {
-	var uploadBundleCmd = &cobra.Command{
-		Use:   "upload-bundle",
-		Short: "Upload plugin bundle to the local system",
-		Long:  "Upload plugin bundle to the local system",
-		RunE: func(cmd *cobra.Command, args []string) error {
-			options := airgapped.UploadPluginBundleOptions{
-				Tar:             upbo.sourceTar,
-				DestinationRepo: upbo.destinationRepo,
-				ImageProcessor:  carvelhelpers.NewImageOperationsImpl(),
-			}
-			return options.UploadPluginBundle()
-		},
-	}
-
-	f := uploadBundleCmd.Flags()
-	f.StringVarP(&upbo.sourceTar, "tar", "", "", "source tar file")
-	f.StringVarP(&upbo.destinationRepo, "to-repo", "", "", "destination repository for publishing plugins")
-
-	return uploadBundleCmd
-}
